+++ CHANGELOG.md
@@ -10,6 +10,11 @@ Note that Sockeye has checks in place to not translate with an old model that wa
 
 Each version section may have have subsections for: _Added_, _Changed_, _Removed_, _Deprecated_, and _Fixed_.
 
+## [1.18.23]
+### Added
+- Now supports negative constraints, which are phrases that must *not* appear in the output.
+  This is achieved with a 'avoid' keyword in the JSON object, with a list of strings as its field value.
+
 ## [1.18.22]
 ### Fixed
 - Make sure the default bucket is large enough with word based batching when the source is longer than the target (Previously
@@ -17,7 +22,7 @@ there was an edge case where the memory usage was sub-optimal with word based ba
 
 ## [1.18.21]
 ### Fixed
-- Constrained decoding was missed a crucial cast
+- Constrained decoding was missing a crucial cast
 - Fixed test cases that should have caught this
 
 ## [1.18.20]